+++ src/main/java/com/ing/data/cassandra/jdbc/commands/CopyFromCommandExecutor.java
@@ -26,7 +26,7 @@ import com.opencsv.exceptions.CsvException;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.IOException;