+++ src/main/java/com/ing/data/cassandra/jdbc/commands/CopyToCommandExecutor.java
@@ -25,7 +25,7 @@ import com.opencsv.ResultSetHelperService;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.nio.charset.StandardCharsets;