+++ src/main/java/com/ing/data/cassandra/jdbc/commands/SpecialCommandsUtil.java
@@ -12,7 +12,7 @@ import com.ing.data.cassandra.jdbc.ColumnDefinitions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 import java.io.File;
 import java.nio.ByteBuffer;
 import java.sql.SQLSyntaxErrorException;