+++ src/main/java/com/ing/data/cassandra/jdbc/types/JdbcTimeUUID.java
@@ -16,7 +16,7 @@
 package com.ing.data.cassandra.jdbc.types;
 
 
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 import java.util.UUID;
 
 /**