+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Update Semver4j to version 6.0.0.
 - Update Astra SDK to 2.1.2.
 - Update AWS Secrets Manager SDK to 2.37.2.
+- Replace Javax JSR-305 dependency by Jakarta Annotations™ API 3.0.0.
 
 ## [4.16.1] - 2025-10-25
 ### Changed